requirements.txt
Flask==3.0.3
Flask-Migrate==4.0.4
Flask-Login==0.6.2
Flask-WTF==1.1.1
Flask-SQLAlchemy==3.0.3
python-telegram-bot==20.3
python-dotenv==1.0.0
Werkzeug==3.0.6
aiohttp==3.10.10
hypercorn==0.16.2
